payments logic & persistence, kafka support & outbox pattern

+++ commons/src/main/java/com/example/commons/dto/OrderDTO.java
@@ -1,20 +1,20 @@
 package com.example.commons.dto;
 
 public class OrderDTO {
-    private Integer id;
+    private Long id;
     private String product;
     private int quantity;
 
     public OrderDTO() {}
 
-    public OrderDTO(Integer id, String product, int quantity) {
+    public OrderDTO(Long id, String product, int quantity) {
         this.id = id;
         this.product = product;
         this.quantity = quantity;
     }
 
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getProduct() { return product; }
     public void setProduct(String product) { this.product = product; }
